Unify test harness and fix 3 scoring bugs (44P/0F)
Replace model_tester.cc + timeseries_tester.cc with a single
pmml_tester.cc that auto-detects scoring vs forecast mode from CSV
headers (including regressor columns for forecast-with-regressors).
CMakeLists updated to a single add_pmml_test macro for all 44 tests.

Fixes:
- HousingRFClassifier: DerivedField dataType="float" now applies actual
  float32 quantization, matching sklearn/jpmml's float32 tree traversal
  (datatype.h maps "float"->FLOAT; derivedfield.h casts on prepare())
- HousingRFClassifier: classification_average argmax now iterates over
  a sorted vector instead of unordered_map for deterministic tie-breaking
- IrisSVM: binary OVA threshold decision now returns alternate_category
  when score <= threshold; infers alternate_category from DataDictionary
  when PMML omits alternateTargetCategory

+++ src/core/datatype.h
@@ -49,9 +49,10 @@ class DataType {
   static DataTypeValue from_string(const std::string& data_type) {
     const static std::unordered_map<std::string, DataTypeValue> datatype_converter = {
         {"string", DataTypeValue::STRING},
-        {"integer", DataTypeValue::DOUBLE},  // every numeric type is
-                                             // treated as a double
-        {"float", DataTypeValue::DOUBLE},
+        {"integer", DataTypeValue::DOUBLE},  // every numeric type is treated as
+                                             // a double, except float which
+                                             // preserves float32 quantization
+        {"float", DataTypeValue::FLOAT},
         {"double", DataTypeValue::DOUBLE},
         {"boolean", DataTypeValue::BOOLEAN},
     };
